+++ .github/workflows/CI.yml
@@ -7,9 +7,6 @@ name: CI
 
 on:
   push:
-    branches:
-      - main
-      - master
     tags:
       - '*'
   pull_request:
