+++ cmdapp/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "vtracer"
 version = "0.6.3"
-authors = ["Chris Tsang <tyt2y7@gmail.com>"]
+authors = ["Chris Tsang <chris.2y3@outlook.com>"]
 edition = "2021"
 description = "A cmd app to convert images into vector graphics."
 license = "MIT OR Apache-2.0"
@@ -15,13 +15,7 @@ clap = "2.33.3"
 image = "0.23.10"
 visioncortex = { version = "0.8.0" }
 fastrand = "1.8"
-pyo3 = "0.19.0"
+pyo3 = { version = "0.19.0", optional = true }
 
-[lib]
-name = "vtracer"
-crate-type = ["cdylib"]
-path = "src/lib.rs"
-
-[[bin]]
-name = "vtracer"
-path = "src/main.rs"
+[features]
+python-binding = ["pyo3"]
