+++ cmdapp/pyproject.toml
@@ -2,8 +2,8 @@
 name = "vtracer"
 version = "0.6.3"
 description = "Python bindings for the Rust Vtracer raster-to-vector library"
-authors = [ { name = "Chris Tsang",  email = "tyt2y7@gmail.com" } ]
-readme = "vtracer/README_PY.md"
+authors = [ { name = "Chris Tsang",  email = "chris.2y3@outlook.com" } ]
+readme = "vtracer/README.md"
 requires-python = ">=3.7"
 license = "MIT"
 classifiers = [
@@ -25,4 +25,4 @@ build-backend = "maturin"
 [tool.maturin]
 features = ["pyo3/extension-module"]
 compatibility = "linux"
-sdist-include = ["../LICENSE", "../README_PY.md"]
+sdist-include = ["../LICENSE", "../README.md"]
