+++ cmdapp/src/main.rs
@@ -1,10 +1,8 @@
-mod config;
-mod converter;
-mod svg;
+use vtracer::{Config, convert_image_to_svg};
 
 fn main() {
-    let config = config::Config::from_args();
-    let result = converter::convert_image_to_svg(config);
+    let config = Config::from_args();
+    let result = convert_image_to_svg(config);
     match result {
         Ok(()) => {
             println!("Conversion successful.");
